requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
biopython>=1.79
tqdm>=4.62.0
pandas>=1.3.0
argparse
dataclasses
typing-extensions>=4.0.0 